refactor: domain 구조 변경
