+++ src/main/java/com/codeit/sprint/team3/backend/bookclub/adapter/in/web/request/CreateBookClubRequest.java
@@ -2,7 +2,7 @@ package com.codeit.sprint.team3.backend.bookclub.adapter.in.web.request;
 
 import com.codeit.sprint.team3.backend.bookclub.domain.BookClubType;
 import com.codeit.sprint.team3.backend.bookclub.domain.MeetingType;
-import com.codeit.sprint.team3.backend.bookclub.domain.model.BookClub;
+import com.codeit.sprint.team3.backend.bookclub.domain.BookClub;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
