+++ src/main/java/com/codeit/sprint/team3/backend/bookclub/adapter/in/web/response/BookClubResponse.java
@@ -2,7 +2,7 @@ package com.codeit.sprint.team3.backend.bookclub.adapter.in.web.response;
 
 import com.codeit.sprint.team3.backend.bookclub.domain.BookClubType;
 import com.codeit.sprint.team3.backend.bookclub.domain.MeetingType;
-import com.codeit.sprint.team3.backend.bookclub.domain.model.BookClub;
+import com.codeit.sprint.team3.backend.bookclub.domain.BookClub;
 
 public record BookClubResponse(
         Long id,
