+++ src/main/java/com/codeit/sprint/team3/backend/bookclub/application/port/in/BookClubMemberUseCase.java
@@ -1,6 +1,6 @@
 package com.codeit.sprint.team3.backend.bookclub.application.port.in;
 
-import com.codeit.sprint.team3.backend.bookclub.domain.model.Member;
+import com.codeit.sprint.team3.backend.bookclub.domain.Member;
 
 public interface BookClubMemberUseCase {
     void saveMember(Member member);
