+++ src/main/java/com/codeit/sprint/team3/backend/bookclub/application/service/BookClubMemberService.java
@@ -3,7 +3,7 @@ package com.codeit.sprint.team3.backend.bookclub.application.service;
 import com.codeit.sprint.team3.backend.bookclub.adapter.out.persistence.entity.BookClubMemberEntity;
 import com.codeit.sprint.team3.backend.bookclub.application.port.out.CommandBookClubMemberPort;
 import com.codeit.sprint.team3.backend.bookclub.application.port.in.BookClubMemberUseCase;
-import com.codeit.sprint.team3.backend.bookclub.domain.model.Member;
+import com.codeit.sprint.team3.backend.bookclub.domain.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
