--- src/main/java/com/codeit/sprint/team3/backend/bookclub/domain/event/BookClubMemberJoinedEvent.java
@@ -1,5 +1,0 @@
-package com.codeit.sprint.team3.backend.bookclub.domain.event;
-
-public class BookClubMemberJoinedEvent {
-    //이벤트 리스너를 통해 모임채팅에 멤버의 참여를 알린다.
-}
